# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.ErrorResponse:
    properties:
      errorCode:
        type: string
      message:
        type: string
    type: object
...
  handler.UpdateItemRequest:
    properties:
      categoryId:
        type: integer
      productId:
        type: integer
      quantity:
        type: number
      total:
        type: number
      unitPrice:
        type: number
    type: object
...
  schemas.CategorySimple:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  schemas.ProductSimple:
    properties:
      id:
        type: integer
      name:
        type: string
      unity:
        type: string
    type: object
...
  schemas.ReceiptItemResponse:
    properties:
      category:
        allOf:
        - $ref: '#/definitions/schemas.CategorySimple'
        description: Apenas ID e Nome
      categoryId:
        type: integer
      createdAt:
        type: string
      id:
        type: integer
      product:
        allOf:
        - $ref: '#/definitions/schemas.ProductSimple'
        description: Nome e unidade do produto
      productId:
        type: integer
      quantity:
        type: number
      receiptId:
        type: integer
      total:
        type: number
      unitPrice:
        type: number
      updatedAt:
        type: string
    type: object
...
host: 147.185.221.212:61489
info:
  contact:
    email: suporte@exemplo.com
    name: Equipe de Desenvolvimento
    url: https://github.com/Pmmvito/Golang-Api-Exemple
  description: "API REST para autenticação JWT, análise inteligente de notas fiscais
    de supermercado e gestão de categorias\n\n## \U0001F680 Funcionalidades Principais:\n-
    \U0001F510 **Autenticação JWT** com tokens de 7 dias\n- \U0001F916 **Análise de
    Notas com IA** usando Google Gemini (2-3s)\n- \U0001F4CA **Categorização Automática**
    de produtos de supermercado\n- \U0001F4DD **CRUD Completo** de recibos e categorias\n-
    \U0001F50D **Filtros Avançados** por categoria, data e valor\n- ⚡ **Respostas
    Otimizadas** (55% menores)\n\n## \U0001F4D6 Como Usar:\n1. Registre-se em `/register`\n2.
    Faça login em `/login` para obter o token JWT\n3. Clique em **Authorize** \U0001F513
    (cadeado verde) e cole: `Bearer SEU_TOKEN`\n4. Agora você pode testar todos os
    endpoints protegidos!\n\n## \U0001F6D2 Categorias de Supermercado (21 categorias
    distintas):\n**Básicos:** Grãos e Cereais, Massas, Padaria\n**Proteínas:** Carnes
    e Proteínas, Frios e Embutidos\n**Laticínios:** Leite, Queijos, Iogurtes, Manteiga\n**Frescos:**
    Frutas e Vegetais\n**Bebidas:** Bebidas (não alcoólicas), Bebidas Alcoólicas,
    Café e Chá\n**Congelados:** Produtos Congelados\n**Doces:** Doces e Sobremesas,
...
    Higiene Pessoal, Limpeza Doméstica, Papel e Descartáveis\n**Especiais:** Bebê
...
paths:
...
  /item/{id}:
    delete:
      consumes:
      - application/json
      - application/json
      description: |-
        Delete an existing receipt item by its ID.
        Soft delete a receipt item (sets deleted_at timestamp). Note: The associated product is NOT deleted as it may be referenced by other items.
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Receipt Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: Item deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized - Invalid or missing token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Item not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Delete receipt item
      tags:
      - items
      - items
    get:
      description: Busca um item pelo ID
      parameters:
      - description: ID do item
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.ReceiptItemResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Buscar item por ID
      tags:
      - items
    patch:
      consumes:
      - application/json
      description: Update an existing receipt item. All fields are optional.
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Item data to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.ReceiptItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an item
      tags:
      - items
  /items:
    get:
      description: Lista todos os itens de recibos do usuário autenticado
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schemas.ReceiptItemResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Listar todos os itens
      tags:
      - items
  /items/date/{date}:
    get:
      description: Lista itens de recibos do usuário autenticado por data de recibo
      parameters:
      - description: Data (YYYY-MM-DD)
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schemas.ReceiptItemResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Listar itens por data de recibo
      tags:
      - items
  /items/period:
    get:
      description: Lista itens de recibos do usuário autenticado entre query params
        `start` e `end` (RFC3339 ou YYYY-MM-DD). Ambos obrigatórios.
      parameters:
      - description: Data/hora inicial (RFC3339 ou YYYY-MM-DD)
        in: query
        name: start
        required: true
        type: string
      - description: Data/hora final (RFC3339 ou YYYY-MM-DD)
        in: query
        name: end
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schemas.ReceiptItemResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Listar itens por período
      tags:
      - items
...
securityDefinitions:
  BearerAuth:
    description: 'Digite "Bearer" seguido do seu token JWT (obtido no login). Exemplo:
      Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"